Simplify schema generator script

+++ schemas-extractor/build-all.sh
@@ -2,6 +2,11 @@
 
 CUR="$(pwd)"
 
+if [ ! -f 'providers.list.full' ]; then
+  echo 'providers.list.full required'
+  exit 1
+fi
+
 out="$CUR/schemas"
 mkdir -p "$out"
 rm -f "$CUR/failure.txt"
@@ -47,10 +52,10 @@ for p in $(cat "$CUR/providers.list.full"); do
   rm -rf generate-schema
   mkdir generate-schema
   cp -r "$CUR/template/generate-schema.go" generate-schema/generate-schema.go
-  find generate-schema -type f -exec sed -i "s/__FULL_NAME__/$p/g" {} +
-  find generate-schema -type f -exec sed -i "s/__NAME__/${p:19}/g" {} +
-  find generate-schema -type f -exec sed -i "s/__REVISION__/$revision/g" {} +
-  find generate-schema -type f -exec sed -i "s~__OUT__~$out~g" {} +
+  sed -i "s/__FULL_NAME__/$p/g" generate-schema/generate-schema.go
+  sed -i "s/__NAME__/${p:19}/g" generate-schema/generate-schema.go
+  sed -i "s/__REVISION__/$revision/g" generate-schema/generate-schema.go
+  sed -i "s~__OUT__~$out~g" generate-schema/generate-schema.go
 
   #echo "Building $p"
   #make
